# project-be-books/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Book reviews",
    "description": "Search Project Gutenberg books and review them. Reviews are accepted right away and enriched in the background with the data of the book.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/review": {
      "post": {
        "tags": [
          "reviews"
        ],
        "summary": "Submit Review",
        "operationId": "submit_review_review_post",
        "security": [
          {
            "ApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^\"?[A-Za-z0-9._:-]{1,255}\"?$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Unique value chosen by the client, such as a UUID. Repeating the request with the same key returns the review created the first time instead of a new one. Keys are remembered for a limited time, 24 hours by default.",
              "title": "Idempotency-Key"
            },
            "description": "Unique value chosen by the client, such as a UUID. Repeating the request with the same key returns the review created the first time instead of a new one. Keys are remembered for a limited time, 24 hours by default."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewSubmission"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            },
            "headers": {
              "ETag": {
                "description": "Entity tag of the review, for If-Match and If-None-Match",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "The X-API-Key header is missing or not valid",
            "headers": {
              "WWW-Authenticate": {
                "description": "Always `ApiKey`",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "413": {
            "description": "The request body is larger than the limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "The request is not valid; `errors` lists the invalid fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "502": {
            "description": "The book catalog failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "A backing service is temporarily unavailable",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before trying again",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "504": {
            "description": "The book catalog did not answer in time",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/review/{review_id}": {
      "get": {
        "tags": [
          "reviews"
        ],
        "summary": "Get Review",
        "operationId": "get_review_review__review_id__get",
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Review Id"
            }
          },
          {
            "name": "if-none-match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Entity tags the client already has: 304 if the review still has one.",
              "title": "If-None-Match"
            },
            "description": "Entity tags the client already has: 304 if the review still has one."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            },
            "headers": {
              "ETag": {
                "description": "Entity tag of the review, for If-Match and If-None-Match",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "202": {
            "description": "The review is saved and still being processed",
            "headers": {
              "ETag": {
                "description": "Entity tag of the review, for If-Match and If-None-Match",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          },
          "304": {
            "description": "The review still has one of the entity tags in If-None-Match"
          },
          "404": {
            "description": "No such review",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "The request is not valid; `errors` lists the invalid fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "A backing service is temporarily unavailable",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before trying again",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "reviews"
        ],
        "summary": "Update Review",
        "operationId": "update_review_review__review_id__put",
        "security": [
          {
            "ApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Review Id"
            }
          },
          {
            "name": "if-match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Entity tag from an earlier response: the review is changed only if it still has it, otherwise the answer is 412.",
              "title": "If-Match"
            },
            "description": "Entity tag from an earlier response: the review is changed only if it still has it, otherwise the answer is 412."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewChanges"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            },
            "headers": {
              "ETag": {
                "description": "Entity tag of the review, for If-Match and If-None-Match",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "The X-API-Key header is missing or not valid",
            "headers": {
              "WWW-Authenticate": {
                "description": "Always `ApiKey`",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "The review belongs to another client",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "412": {
            "description": "The review no longer has the entity tag given in If-Match",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "413": {
            "description": "The request body is larger than the limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "No such review",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "The request is not valid; `errors` lists the invalid fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "A backing service is temporarily unavailable",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before trying again",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "reviews"
        ],
        "summary": "Delete Review",
        "operationId": "delete_review_review__review_id__delete",
        "security": [
          {
            "ApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Review Id"
            }
          },
          {
            "name": "if-match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Entity tag from an earlier response: the review is changed only if it still has it, otherwise the answer is 412.",
              "title": "If-Match"
            },
            "description": "Entity tag from an earlier response: the review is changed only if it still has it, otherwise the answer is 412."
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "The X-API-Key header is missing or not valid",
            "headers": {
              "WWW-Authenticate": {
                "description": "Always `ApiKey`",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "The review belongs to another client",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "412": {
            "description": "The review no longer has the entity tag given in If-Match",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "No such review",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "The request is not valid; `errors` lists the invalid fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "A backing service is temporarily unavailable",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before trying again",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuthorDetails": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "birth_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Year"
          },
          "death_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Death Year"
          }
        },
        "type": "object",
        "required": [
          "name",
          "birth_year",
          "death_year"
        ],
        "title": "AuthorDetails"
      },
      "BookDetails": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "authors": {
            "items": {
              "$ref": "#/components/schemas/AuthorDetails"
            },
            "type": "array",
            "title": "Authors"
          },
          "subjects": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Subjects"
          },
          "bookshelves": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Bookshelves"
          },
          "languages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Languages"
          },
          "summaries": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Summaries"
          },
          "cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          },
          "download_count": {
            "type": "integer",
            "title": "Download Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "authors",
          "subjects",
          "bookshelves",
          "languages",
          "summaries",
          "cover_url",
          "download_count"
        ],
        "title": "BookDetails"
      },
...
      "FieldError": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "title": "FieldError"
      },
      "ProblemDetails": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "instance": {
            "type": "string",
            "title": "Instance"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "title",
          "status",
          "instance"
        ],
        "title": "ProblemDetails"
      },
      "ReviewChanges": {
        "properties": {
          "review": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 3,
            "title": "Review",
            "description": "Text of the review; line breaks and tabs are allowed."
          },
          "score": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Score"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "review",
          "score"
        ],
        "title": "ReviewChanges"
      },
      "ReviewResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/ReviewStatus"
          },
          "book_id": {
            "type": "integer",
            "title": "Book Id"
          },
          "review": {
            "type": "string",
            "title": "Review"
          },
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "book": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BookDetails"
              },
              {
                "type": "null"
              }
            ],
            "description": "Book data from the catalog, available once the review is completed."
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "book_id",
          "review",
          "score",
          "created_at",
          "updated_at",
          "book"
        ],
        "title": "ReviewResponse"
      },
      "ReviewStatus": {
        "type": "string",
        "enum": [
          "pending",
          "completed",
          "failed"
        ],
        "title": "ReviewStatus"
      },
      "ReviewSubmission": {
        "properties": {
          "id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Id",
            "description": "Gutenberg ID of the book, as returned by /book/search."
          },
          "review": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 3,
            "title": "Review",
            "description": "Text of the review; line breaks and tabs are allowed."
          },
          "score": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Score"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "review",
          "score"
        ],
        "title": "ReviewSubmission"
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "Key of the client application, required to create, change and delete reviews.",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}